move task to common dir
