+++ BUILD.bazel
@@ -183,6 +183,32 @@ cc_library(
 
 # === End of rpc definitions ===
 
+cc_library(
+    name = "ray_common",
+    srcs = glob(
+        [
+            "src/ray/common/*.cc",
+        ],
+        exclude = [
+            "src/ray/common/*_test.cc",
+        ],
+    ),
+    hdrs = glob(
+        [
+            "src/ray/common/*.h",
+        ],
+    ),
+    copts = COPTS,
+    deps = [
+        ":common_cc_proto",
+        ":ray_util",
+        ":node_manager_fbs",
+        "@boost//:asio",
+        "@com_github_grpc_grpc//:grpc++",
+        "@plasma//:plasma_client",
+    ],
+)
+
 cc_binary(
     name = "raylet",
     srcs = ["src/ray/raylet/main.cc"],
@@ -491,29 +517,6 @@ cc_library(
     deps = [
         ":sha256",
         "@com_github_google_glog//:glog",
-        "@plasma//:plasma_client",
-    ],
-)
-
-cc_library(
-    name = "ray_common",
-    srcs = glob(
-        [
-            "src/ray/common/*.cc",
-        ],
-        exclude = [
-            "src/ray/common/*_test.cc",
-        ],
-    ),
-    hdrs = glob(
-        [
-            "src/ray/common/*.h",
-        ],
-    ),
-    copts = COPTS,
-    deps = [
-        ":ray_util",
-        "@boost//:asio",
         "@plasma//:plasma_client",
     ],
 )
