+++ src/ray/common/task_common.h
@@ -1,0 +1,13 @@
+#ifndef RAY_COMMON_TASK_COMMON_H
+#define RAY_COMMON_TASK_COMMON_H
+
+#include "ray/protobuf/common.pb.h"
+
+namespace ray {
+
+using rpc::Language;
+using rpc::TaskType;
+
+}
+
+#endif
