+++ src/ray/core_worker/context.h
@@ -2,7 +2,7 @@
 #define RAY_CORE_WORKER_CONTEXT_H
 
 #include "ray/core_worker/common.h"
-#include "ray/raylet/task_spec.h"
+#include "ray/common/task_spec.h"
 
 namespace ray {
