+++ src/ray/gcs/tables.cc
@@ -3,7 +3,7 @@
 #include "ray/common/common_protocol.h"
 #include "ray/common/ray_config.h"
 #include "ray/gcs/client.h"
-#include "ray/rpc/util.h"
+#include "ray/common/grpc_util.h"
 #include "ray/util/util.h"
 
 namespace {
