+++ src/ray/raylet/actor_registration.h
@@ -5,7 +5,7 @@
 
 #include "ray/common/id.h"
 #include "ray/protobuf/gcs.pb.h"
-#include "ray/raylet/task.h"
+#include "ray/common/task.h"
 
 namespace ray {
