+++ src/ray/raylet/lineage_cache.h
@@ -7,7 +7,7 @@
 #include "ray/common/id.h"
 #include "ray/common/status.h"
 #include "ray/gcs/tables.h"
-#include "ray/raylet/task.h"
+#include "ray/common/task.h"
 
 namespace ray {
