+++ src/ray/raylet/node_manager.h
@@ -7,7 +7,7 @@
 #include "ray/rpc/client_call.h"
 #include "ray/rpc/node_manager/node_manager_server.h"
 #include "ray/rpc/node_manager/node_manager_client.h"
-#include "ray/raylet/task.h"
+#include "ray/common/task.h"
 #include "ray/object_manager/object_manager.h"
 #include "ray/common/client_connection.h"
 #include "ray/protobuf/common.pb.h"
@@ -15,7 +15,7 @@
 #include "ray/raylet/lineage_cache.h"
 #include "ray/raylet/scheduling_policy.h"
 #include "ray/raylet/scheduling_queue.h"
-#include "ray/raylet/scheduling_resources.h"
+#include "ray/common/scheduling_resources.h"
 #include "ray/raylet/reconstruction_policy.h"
 #include "ray/raylet/task_dependency_manager.h"
 #include "ray/raylet/worker_pool.h"
