+++ src/ray/raylet/raylet.h
@@ -9,7 +9,7 @@
 // clang-format off
 #include "ray/raylet/node_manager.h"
 #include "ray/object_manager/object_manager.h"
-#include "ray/raylet/scheduling_resources.h"
+#include "ray/common/scheduling_resources.h"
 // clang-format on
 
 namespace ray {
