+++ src/ray/raylet/task_dependency_manager_test.cc
@@ -6,7 +6,7 @@
 #include <boost/asio.hpp>
 
 #include "ray/raylet/task_dependency_manager.h"
-#include "ray/raylet/task_util.h"
+#include "ray/common/task_util.h"
 
 namespace ray {
