+++ src/ray/rpc/server_call.h
@@ -4,7 +4,7 @@
 #include <grpcpp/grpcpp.h>
 
 #include "ray/common/status.h"
-#include "ray/rpc/util.h"
+#include "ray/common/grpc_util.h"
 
 namespace ray {
 namespace rpc {
